Cache similar requests as ids at the model level
The expensive part of the operation (for larger installs) is the
calculation of similar requests in xapian. We were previously
duplicating the meaningful data that was cached by using locale
as a cache key. The similar requests for a request are identical
over locales. This was causing us to use cache space unnecessarily.
We also were caching the template contents that rendered the
similar requests. Again, this uses cache space on content that can
be fairly cheaply regenerated. Here we cache only the expensive-to-
generate list of ids, which will not take up as much cache room, and
therefore increase the likelihood that any given cache item will
be used before it is evicted.